chat: extract AI customization item sources
